+++ src/realm/table_view.cpp
@@ -670,10 +670,23 @@ void TableViewBase::sync_distinct_view(size_t column)
     m_distinct_column_source = column;
     if (m_distinct_column_source != npos) {
         REALM_ASSERT(m_table);
-        REALM_ASSERT(m_table->has_search_index(m_distinct_column_source));
         if (!m_table->is_degenerate()) {
             const ColumnBase& col = m_table->get_column_base(m_distinct_column_source);
-            col.get_search_index()->distinct(m_row_indexes);
+            if (m_table->has_search_index(m_distinct_column_source)) {
+                col.get_search_index()->distinct(m_row_indexes);
+            }
+            else {
+                m_row_indexes.clear();
+                for (size_t i = 0; i < m_table->size(); i++)
+                    m_row_indexes.add(i);
+
+                // Returns results in sorted order, to be consistent with StringIndex::distinct().
+                // Note that TableViewBase::distinct() guaruantees that results will be returned in
+                // the original column order (not necessarily sorted), but Table::get_distinct_view()
+                // makes no promise about the order of results and is currently returning sorted results.
+                sort(column);
+                do_get_distinct(std::vector<size_t> {column});
+            }
         }
     }
 }
@@ -694,6 +707,11 @@ void TableViewBase::distinct(std::vector<size_t> columns)
     if (m_distinct_columns.size() == 0)
         return;
 
+    do_get_distinct(m_distinct_columns);
+}
+
+void TableViewBase::do_get_distinct(std::vector<size_t> columns)
+{
     // Step 1: First copy original TableView into a vector
     std::vector<size_t> original;
     original.resize(size());
@@ -716,7 +734,7 @@ void TableViewBase::distinct(std::vector<size_t> columns)
     m_columns_enum.resize(columns.size());
 
     for (size_t i = 0; i < columns.size(); i++) {
-        const ColumnBase& cb = m_table->get_column_base(m_distinct_columns[i]);
+        const ColumnBase& cb = m_table->get_column_base(columns[i]);
         // FIXME: If we decide to keep StringEnumColumn (see Table::optimize()), then below conditional type casting
         // should be removed in favor for a more elegant/generalized solution, because this casting pattern is used
         // in a couple of other places in Core too.
@@ -734,7 +752,7 @@ void TableViewBase::distinct(std::vector<size_t> columns)
     std::unordered_set<size_t> remove;
     for (size_t r = 1; r < size(); r++) {
         bool identical = true;
-        for (size_t c = 0; c < m_distinct_columns.size(); c++) {
+        for (size_t c = 0; c < columns.size(); c++) {
 
             int cmp;
             int64_t r1_64 = m_row_indexes.get(r);
@@ -767,7 +785,6 @@ void TableViewBase::distinct(std::vector<size_t> columns)
         }
     }
 }
-
 
 // Sort according to one column
 void TableViewBase::sort(size_t column, bool ascending)