+++ test/test_table.cpp
@@ -1503,8 +1503,52 @@ TEST_TYPES(Table_SetStringUnique, StringColumnType, StringEnumColumnType)
 }
 
 
-TEST(Table_Distinct)
+TEST_TYPES(Table_DistinctOnEmptyCol, std::true_type, std::false_type)
 {
+    constexpr bool add_search_index = TEST_TYPE::value;
+
+    Table table;
+    table.add_column(type_Int, "first");
+
+    if (add_search_index) {
+        table.add_search_index(0);
+    }
+    CHECK_EQUAL(table.has_search_index(0), add_search_index);
+
+    TableView view = table.get_distinct_view(0);
+
+    CHECK_EQUAL(0, view.size());
+    CHECK_EQUAL(0, table.size());
+}
+
+
+TEST_TYPES(Table_DistinctWithOneRow, std::true_type, std::false_type)
+{
+    constexpr bool add_search_index = TEST_TYPE::value;
+
+    Table table;
+    table.add_column(type_Int, "first");
+
+    if (add_search_index) {
+        table.add_search_index(0);
+    }
+    CHECK_EQUAL(table.has_search_index(0), add_search_index);
+    table.add_empty_row();
+    CHECK_EQUAL(1, table.size());
+    table.set_int(0, 0, 42);
+    CHECK_EQUAL(42, table.get_int(0, 0));
+
+    TableView view = table.get_distinct_view(0);
+
+    CHECK_EQUAL(1, view.size());
+    CHECK_EQUAL(0, view.get_source_ndx(0));
+}
+
+
+TEST_TYPES(Table_Distinct, std::true_type, std::false_type)
+{
+    constexpr bool add_search_index = TEST_TYPE::value;
+
     TestTableEnum table;
 
     table.add(Mon, "A");
@@ -1516,8 +1560,11 @@ TEST(Table_Distinct)
     table.add(Sun, "D");
     table.add(Mon, "D");
 
-    table.column().second.add_search_index();
-    CHECK(table.column().second.has_search_index());
+    if (add_search_index) {
+        table.column().second.add_search_index();
+    }
+    CHECK_EQUAL(table.column().second.has_search_index(), add_search_index);
+
 
     TestTableEnum::View view = table.column().second.get_distinct_view();
 
@@ -1529,8 +1576,10 @@ TEST(Table_Distinct)
 }
 
 
-TEST(Table_DistinctEnums)
+TEST_TYPES(Table_DistinctEnums, std::true_type, std::false_type)
 {
+    constexpr bool add_search_index = TEST_TYPE::value;
+
     TestTableEnum table;
     table.add(Mon, "A");
     table.add(Tue, "B");
@@ -1541,8 +1590,10 @@ TEST(Table_DistinctEnums)
     table.add(Sun, "D");
     table.add(Mon, "D");
 
-    table.column().first.add_search_index();
-    CHECK(table.column().first.has_search_index());
+    if (add_search_index) {
+        table.column().first.add_search_index();
+    }
+    CHECK_EQUAL(table.column().first.has_search_index(), add_search_index);
 
     TestTableEnum::View view = table.column().first.get_distinct_view();
 
@@ -1557,8 +1608,10 @@ TEST(Table_DistinctEnums)
 }
 
 
-TEST(Table_DistinctIntegers)
+TEST_TYPES(Table_DistinctIntegers, std::true_type, std::false_type)
 {
+    constexpr bool add_search_index = TEST_TYPE::value;
+
     Table table;
     table.add_column(type_Int, "first");
     table.add_empty_row(4);
@@ -1567,8 +1620,10 @@ TEST(Table_DistinctIntegers)
     table.set_int(0, 2, 3);
     table.set_int(0, 3, 3);
 
-    table.add_search_index(0);
-    CHECK(table.has_search_index(0));
+    if (add_search_index) {
+        table.add_search_index(0);
+    }
+    CHECK_EQUAL(table.has_search_index(0), add_search_index);
 
     TableView view = table.get_distinct_view(0);
 
@@ -1579,8 +1634,10 @@ TEST(Table_DistinctIntegers)
 }
 
 
-TEST(Table_DistinctBool)
+TEST_TYPES(Table_DistinctBool, std::true_type, std::false_type)
 {
+    constexpr bool add_search_index = TEST_TYPE::value;
+
     Table table;
     table.add_column(type_Bool, "first");
     table.add_empty_row(4);
@@ -1589,8 +1646,10 @@ TEST(Table_DistinctBool)
     table.set_bool(0, 2, true);
     table.set_bool(0, 3, false);
 
-    table.add_search_index(0);
-    CHECK(table.has_search_index(0));
+    if (add_search_index) {
+        table.add_search_index(0);
+    }
+    CHECK_EQUAL(table.has_search_index(0), add_search_index);
 
     TableView view = table.get_distinct_view(0);
 
@@ -1600,25 +1659,25 @@ TEST(Table_DistinctBool)
 }
 
 
-/*
-// FIXME Commented out because indexes on floats and doubles are not supported (yet).
-
 TEST(Table_DistinctFloat)
 {
     Table table;
-    table.add_column(type_Float, "first");
-    table.add_empty_row(12);
+    table.add_column(type_Float, "first", true);
+    table.add_empty_row(14);
     for (size_t i = 0; i < 10; ++i) {
         table.set_float(0, i, static_cast<float>(i) + 0.5f);
     }
     table.set_float(0, 10, 0.5f);
     table.set_float(0, 11, 1.5f);
+    table.set_null(0, 12);
+    table.set_null(0, 13);
 
-    table.add_search_index(0);
-    CHECK(table.has_search_index(0));
+    // Search index not supported for float
+    //table.add_search_index(0);
+    //CHECK(table.has_search_index(0));
 
     TableView view = table.get_distinct_view(0);
-    CHECK_EQUAL(10, view.size());
+    CHECK_EQUAL(11, view.size());
 }
 
 
@@ -1633,17 +1692,39 @@ TEST(Table_DistinctDouble)
     table.set_double(0, 10, 0.5);
     table.set_double(0, 11, 1.5);
 
-    table.add_search_index(0);
-    CHECK(table.has_search_index(0));
+    // Search index not supported for double
+    //table.add_search_index(0);
+    //CHECK(table.has_search_index(0));
 
     TableView view = table.get_distinct_view(0);
     CHECK_EQUAL(10, view.size());
 }
-*/
 
 
-TEST(Table_DistinctDateTime)
+TEST(Table_DistinctBinary)
 {
+    Table table;
+    table.add_column(type_Binary, "first");
+    table.add_empty_row(5);
+    table.set_binary(0, 0, BinaryData("aaa"));
+    table.set_binary(0, 1, BinaryData("bbb"));
+    table.set_binary(0, 2, BinaryData("ccc"));
+    table.set_binary(0, 3, BinaryData("aaa"));
+    table.set_binary(0, 4, BinaryData("zzz"));
+
+    TableView view = table.get_distinct_view(0);
+    CHECK_EQUAL(4, view.size());
+    CHECK_EQUAL(0, view.get_source_ndx(0));
+    CHECK_EQUAL(1, view.get_source_ndx(1));
+    CHECK_EQUAL(2, view.get_source_ndx(2));
+    CHECK_EQUAL(4, view.get_source_ndx(3));
+}
+
+
+TEST_TYPES(Table_DistinctDateTime, std::true_type, std::false_type)
+{
+    constexpr bool add_search_index = TEST_TYPE::value;
+
     Table table;
     table.add_column(type_OldDateTime, "first");
     table.add_empty_row(4);
@@ -1652,16 +1733,111 @@ TEST(Table_DistinctDateTime)
     table.set_olddatetime(0, 2, OldDateTime(3));
     table.set_olddatetime(0, 3, OldDateTime(3));
 
-    table.add_search_index(0);
-    CHECK(table.has_search_index(0));
+    if (add_search_index) {
+        table.add_search_index(0);
+    }
+    CHECK_EQUAL(table.has_search_index(0), add_search_index);
 
     TableView view = table.get_distinct_view(0);
     CHECK_EQUAL(3, view.size());
 }
 
-
-TEST(Table_DistinctFromPersistedTable)
+TEST_TYPES(Table_DistinctTimestamp, std::true_type, std::false_type)
 {
+    constexpr bool add_search_index = TEST_TYPE::value;
+
+    Table table;
+    table.add_column(type_Timestamp, "first");
+    table.add_empty_row(4);
+    table.set_timestamp(0, 0, Timestamp(3, 3));
+    table.set_timestamp(0, 1, Timestamp(0, 0));
+    table.set_timestamp(0, 2, Timestamp(1, 1));
+    table.set_timestamp(0, 3, Timestamp(3, 3));
+
+    if (add_search_index) {
+        table.add_search_index(0);
+    }
+    CHECK_EQUAL(table.has_search_index(0), add_search_index);
+
+    TableView view = table.get_distinct_view(0);
+    CHECK_EQUAL(3, view.size());
+    CHECK_EQUAL(1, view.get_source_ndx(0));
+    CHECK_EQUAL(2, view.get_source_ndx(1));
+    CHECK_EQUAL(0, view.get_source_ndx(2));
+}
+
+TEST_TYPES(Table_DistinctStringEnum, std::true_type, std::false_type)
+{
+    constexpr bool add_search_index = TEST_TYPE::value;
+
+    Table table;
+    table.add_column(type_String, "first");
+    table.add_empty_row(4);
+    table.set_string(0, 0, "a");
+    table.set_string(0, 1, "b");
+    table.set_string(0, 2, "c");
+    table.set_string(0, 3, "c");
+
+    if (add_search_index) {
+        table.add_search_index(0);
+    }
+    CHECK_EQUAL(table.has_search_index(0), add_search_index);
+    table.optimize(true);   // Force create string enum column
+
+    TableView view = table.get_distinct_view(0);
+    CHECK_EQUAL(3, view.size());
+    CHECK_EQUAL(0, view.get_source_ndx(0));
+    CHECK_EQUAL(1, view.get_source_ndx(1));
+    CHECK_EQUAL(2, view.get_source_ndx(2));
+}
+
+
+TEST(Table_DistinctNull)
+{
+    std::vector<DataType> nullable_column_types {
+        type_Int,
+        type_Bool,
+        type_Float,
+        type_Double,
+        type_String,
+        type_Binary,
+        type_OldDateTime,
+        type_Timestamp
+    };
+
+    Table table;
+    bool nullable = true;
+    size_t num_rows = 5;
+
+    for (DataType t : nullable_column_types) {
+        table.add_column(t, "", nullable);
+    }
+
+    table.add_empty_row(num_rows);
+
+    for (size_t col_ndx = 0; col_ndx < table.get_column_count(); ++ col_ndx) {
+        for (size_t row_ndx = 0; row_ndx < num_rows; ++row_ndx) {
+            table.set_null(col_ndx, row_ndx);
+        }
+        TableView view = table.get_distinct_view(col_ndx);
+        CHECK_EQUAL(1, view.size());
+        CHECK_EQUAL(0, view.get_source_ndx(0));
+    }
+
+    // Test string enum column with null
+    table.optimize(true);
+    auto string_col_pos = 4; // Index of StringColumn in nullable_column_types array
+    CHECK_EQUAL(type_String, table.get_column_type(string_col_pos)); // Hidden StringEnum type
+    TableView view = table.get_distinct_view(string_col_pos);
+    CHECK_EQUAL(1, view.size());
+    CHECK_EQUAL(0, view.get_source_ndx(0));
+}
+
+
+TEST_TYPES(Table_DistinctFromPersistedTable, std::true_type, std::false_type)
+{
+    constexpr bool add_search_index = TEST_TYPE::value;
+
     GROUP_TEST_PATH(path);
 
     {
@@ -1674,8 +1850,10 @@ TEST(Table_DistinctFromPersistedTable)
         table->set_int(0, 2, 3);
         table->set_int(0, 3, 3);
 
-        table->add_search_index(0);
-        CHECK(table->has_search_index(0));
+        if (add_search_index) {
+            table->add_search_index(0);
+        }
+        CHECK_EQUAL(table->has_search_index(0), add_search_index);
         group.write(path);
     }