+++ release_notes.md
@@ -27,7 +27,8 @@
 
 ### Enhancements
 
-* Lorem ipsum.
+* Distinct is now supported for columns without a search index. Bindings no longer
+  need to ensure that a column has a search index before calling distinct. (#1739)
 
 -----------